+++ app/workers/FiscalizationWorker.js
@@ -35,7 +35,7 @@ class FiscalizationWorker {
     }
 
     async processReceipt() {
-        const receipt = await this.receiptService.getFirstPending()
+        const receipt = await this.receiptService.getRandomPending()
 
         if (!receipt) {
             return
@@ -92,9 +92,9 @@ class FiscalizationWorker {
 
             if (e instanceof UmkaResponseError) {
                 const {json} = e
-                const time = await this.cacheService.get(redisProcessingPrefix + receipt.id)
-                const expireDate = new Date(Number(time) + Number(process.env.FISCAL_PENDING_TIMEOUT_SECONDS) * 1000)
-                const expired = (new Date() > expireDate)
+                const createDate = receipt.createdAt
+                const expireDate = new Date(createDate.getTime() + Number(process.env.FISCAL_PENDING_TIMEOUT_SECONDS) * 1000)
+                const expired = (new Date() > expireDate )
 
                 if (expired) {
                     logger.error(`worker_process_receipt_timeout ${receipt.id} ${JSON.stringify(e.json)}`)